third_party/WebKit/LayoutTests/netinfo/estimate-web-worker.html

<!DOCTYPE html>
<head>
<script src="../resources/js-test.js"></script>
<script src="resources/netinfo_common.js"></script>
</head>
<body>
<script>

description('Tests that web-workers have access to NetInfo.');

shouldBe('typeof window.internals.observeGC', '"function"',
'this test requires window.internals');

var worker = new Worker("resources/web-worker.js");

var msg_count = 0;

worker.addEventListener('message', function(e) {
    if (msg_count == 0) {
        if (e.data != connection.type + ',' + connection.downlinkMax + ',' + connection.effectiveType + ',' + connection.rtt + ',' + connection.downlink) {
            testFailed("Worker type disagrees with main frame.");
        }
       internals.setNetworkQualityInfoOverride(newEffectiveType, newRtt, newDownlink);
    } else if (msg_count == 1) {
        if (e.data != connection.type + ',' + connection.downlinkMax + ',' + newEffectiveType + ',' + newRtt + ',' + newDownlink) {
            testFailed("Worker switched to wrong quality estimate.");
        }
        internals.setNetworkQualityInfoOverride(initialEffectiveType, initialRtt, initialDownlink);
    } else if (msg_count == 2) {
        if (e.data != connection.type + ',' + connection.downlinkMax + ',' + initialEffectiveType + ',' + initialRtt + ',' + initialDownlink) {
            testFailed("Worker did not revert back to initial quality estimate.");
        }
        finishJSTest();
    }
    msg_count += 1;
});

worker.postMessage('kickoff');
</script>
</body>
</html>
